VFS name resolver cleanup (#655)
* Separate out (from vfs_open) file creation to vfs_create.
* Move vfs_open to file where it's actually used.
* Move do_* functions operating on files only to file_syscalls.c
