+++ include/sys/vfs.h
@@ -37,6 +37,32 @@ typedef struct componentname {
   size_t cn_namelen;
 } componentname_t;
 
+typedef struct {
+  /*
+   * Arguments fed into name resolver.
+   */
+  vnode_t *vs_atdir; /* startup dir, cwd if null */
+  vnrop_t vs_op;     /* vnr operation type */
+  uint32_t vs_flags; /* flags to vfs name resolver */
+  const char *vs_path;
+
+  /*
+   * Results returned from name resolver.
+   */
+  vnode_t *vs_vp;            /* vnode of result */
+  vnode_t *vs_dvp;           /* vnode of parent directory */
+  componentname_t vs_lastcn; /* last component's name */
+
+  /*
+   * Name resolver internal state. DO NOT TOUCH!
+   */
+  char *vs_pathbuf;  /* pathname buffer */
+  size_t vs_pathlen; /* remaining chars in path */
+  componentname_t vs_cn;
+  const char *vs_nextcn;
+  int vs_loopcnt; /* count of symlinks encountered */
+} vnrstate_t;
+
 #define COMPONENTNAME(str)                                                     \
   (componentname_t) {                                                          \
     .cn_flags = 0, .cn_nameptr = str, .cn_namelen = strlen(str)                \
@@ -44,16 +70,8 @@ typedef struct componentname {
 
 bool componentname_equal(const componentname_t *cn, const char *name);
 
-/* Kernel interface */
+/* Procedures called by system calls implementation. */
 int do_open(proc_t *p, char *pathname, int flags, mode_t mode, int *fd);
-int do_close(proc_t *p, int fd);
-int do_read(proc_t *p, int fd, uio_t *uio);
-int do_write(proc_t *p, int fd, uio_t *uio);
-int do_lseek(proc_t *p, int fd, off_t offset, int whence, off_t *newoffp);
-int do_fstat(proc_t *p, int fd, stat_t *sb);
-int do_dup(proc_t *p, int oldfd, int *newfdp);
-int do_dup2(proc_t *p, int oldfd, int newfd);
-int do_fcntl(proc_t *p, int fd, int cmd, int arg, int *resp);
 int do_unlink(proc_t *p, char *path);
 int do_mkdir(proc_t *p, char *path, mode_t mode);
 int do_rmdir(proc_t *p, char *path);
@@ -66,11 +84,9 @@ int do_symlink(proc_t *p, char *path, char *link);
 ssize_t do_readlinkat(proc_t *p, int fd, char *path, uio_t *uio);
 int do_symlinkat(proc_t *p, char *target, int newdirfd, char *linkpath);
 int do_rename(proc_t *p, char *from, char *to);
-int do_chdir(proc_t *p, char *path);
+int do_chdir(proc_t *p, const char *path);
 int do_fchdir(proc_t *p, int fd);
 int do_getcwd(proc_t *p, char *buf, size_t *lastp);
-int do_umask(proc_t *p, int newmask, int *oldmaskp);
-int do_ioctl(proc_t *p, int fd, u_long cmd, void *data);
 int do_truncate(proc_t *p, char *path, off_t length);
 int do_ftruncate(proc_t *p, int fd, off_t length);
 int do_fstatat(proc_t *p, int fd, char *path, stat_t *sb, int flag);
@@ -80,33 +96,22 @@ int do_mount(const char *fs, const char *path);
 int do_statfs(proc_t *p, char *path, statfs_t *buf);
 int do_getdents(proc_t *p, int fd, uio_t *uio);
 
+/* Initialize & destroy structures required to perform name resolution. */
+int vnrstate_init(vnrstate_t *vs, vnrop_t op, uint32_t flags, const char *path);
+void vnrstate_destroy(vnrstate_t *vs);
+
+/* Perform name resolution with specified operation. */
+int vfs_nameresolve(vnrstate_t *vs);
+
 /* Finds the vnode corresponding to the given path.
  * Increases use count on returned vnode. */
-int vfs_namelookupat(const char *path, vnode_t *atdir, uint32_t flags,
-                     vnode_t **vp);
-#define vfs_namelookup(path, vp) vfs_namelookupat(path, NULL, VNR_FOLLOW, vp)
-
-/* Yield the vnode for an existing entry; or, if there is none, yield NULL.
- * Parent vnode is locked and held; vnode, if exists, is only held.*/
-int vfs_namecreateat(const char *path, vnode_t *atdir, uint32_t flags,
-                     vnode_t **dvp, vnode_t **vp, componentname_t *cn);
-#define vfs_namecreate(path, dvp, vp, cn)                                      \
-  vfs_namecreateat(path, NULL, VNR_FOLLOW, dvp, vp, cn)
-
-/* Both vnode and its parent is held and locked. */
-int vfs_namedeleteat(const char *path, vnode_t *atdir, uint32_t flags,
-                     vnode_t **dvp, vnode_t **vp, componentname_t *cn);
-#define vfs_namedelete(path, dvp, vp, cn)                                      \
-  vfs_namedeleteat(path, NULL, VNR_FOLLOW, dvp, vp, cn)
+int vfs_namelookup(const char *path, vnode_t **vp);
 
 /* Uncovers mountpoint if node is mounted. */
 void vfs_maybe_ascend(vnode_t **vp);
 
 /* Get the root of filesystem if node is a mountpoint. */
 int vfs_maybe_descend(vnode_t **vp);
-
-/* Looks up the vnode corresponding to the pathname and opens it into f. */
-int vfs_open(file_t *f, char *pathname, int flags, int mode);
 
 /* Finds name of v-node in given directory. */
 int vfs_name_in_dir(vnode_t *dv, vnode_t *v, char *buf, size_t *lastp);
