+++ sys/kern/vfs.c
@@ -197,53 +197,4 @@ int vfs_maybe_descend(vnode_t **vp) {
   return 0;
 }
 
-int vfs_open(file_t *f, char *pathname, int flags, int mode) {
-  vnode_t *v;
-  int error = 0;
-  if (flags & O_CREAT) {
-    vnode_t *dvp;
-    componentname_t cn;
-    if ((error = vfs_namecreate(pathname, &dvp, &v, &cn)))
-      return error;
-
-    if (v == NULL) {
-      vattr_t va;
-      vattr_null(&va);
-      va.va_mode = S_IFREG | (mode & ALLPERMS);
-      error = VOP_CREATE(dvp, &cn, &va, &v);
-      vnode_put(dvp);
-      if (error)
-        return error;
-      flags &= ~O_TRUNC;
-    } else {
-      if (v == dvp)
-        vnode_drop(dvp);
-      else
-        vnode_put(dvp);
-
-      if (mode & O_EXCL)
-        error = EEXIST;
-      mode &= ~O_CREAT;
-    }
-  } else {
-    if ((error = vfs_namelookup(pathname, &v)))
-      return error;
-  }
-
-  if (!error && flags & O_TRUNC) {
-    vattr_t va;
-    vattr_null(&va);
-    va.va_size = 0;
-    error = VOP_SETATTR(v, &va);
-  }
-
-  if (!error)
-    error = VOP_OPEN(v, flags, f);
-
-  /* Drop our reference to v. We received it from vfs_namelookup, but we no
-     longer need it - file f keeps its own reference to v after open. */
-  vnode_drop(v);
-  return error;
-}
-
 SYSINIT_ADD(vfs, vfs_init, NODEPS);
