+++ sys/kern/vfs_syscalls.c
@@ -10,8 +10,101 @@
 #include <sys/unistd.h>
 #include <sys/stat.h>
 #include <sys/libkern.h>
-#include <sys/malloc.h>
-#include <sys/mount.h>
+
+static int vfs_nameresolveat(proc_t *p, int fdat, vnrstate_t *vs) {
+  file_t *f;
+  int error;
+
+  if (fdat != AT_FDCWD) {
+    if ((error = fdtab_get_file(p->p_fdtable, fdat, FF_READ, &f)))
+      return error;
+    vs->vs_atdir = f->f_vnode;
+  }
+  error = vfs_nameresolve(vs);
+  if (fdat != AT_FDCWD)
+    file_drop(f);
+
+  return error;
+}
+
+static int vfs_namelookupat(proc_t *p, int fdat, uint32_t flags,
+                            const char *path, vnode_t **vp) {
+  vnrstate_t vs;
+  int error;
+
+  if ((error = vnrstate_init(&vs, VNR_LOOKUP, flags, path)))
+    return error;
+
+  error = vfs_nameresolve(&vs);
+  *vp = vs.vs_vp;
+
+  vnrstate_destroy(&vs);
+  return error;
+}
+
+static int vfs_create(char *pathname, int *flagsp, int mode, vnode_t **vp) {
+  vnrstate_t vs;
+  int error;
+
+  if ((error = vnrstate_init(&vs, VNR_CREATE, VNR_FOLLOW, pathname)))
+    return error;
+
+  if ((error = vfs_nameresolve(&vs)))
+    goto fail;
+
+  if (vs.vs_vp == NULL) {
+    vattr_t va;
+    vattr_null(&va);
+    va.va_mode = S_IFREG | (mode & ALLPERMS);
+    error = VOP_CREATE(vs.vs_dvp, &vs.vs_lastcn, &va, &vs.vs_vp);
+    vnode_put(vs.vs_dvp);
+    if (error)
+      goto fail;
+    *flagsp &= ~O_TRUNC;
+  } else {
+    if (vs.vs_vp == vs.vs_dvp)
+      vnode_drop(vs.vs_dvp);
+    else
+      vnode_put(vs.vs_dvp);
+
+    if (*flagsp & O_EXCL)
+      error = EEXIST;
+    *flagsp &= ~O_CREAT;
+  }
+  *vp = vs.vs_vp;
+
+fail:
+  vnrstate_destroy(&vs);
+  return error;
+}
+
+static int vfs_open(file_t *f, char *pathname, int flags, int mode) {
+  vnode_t *v;
+  int error = 0;
+
+  if (flags & O_CREAT) {
+    if ((error = vfs_create(pathname, &flags, mode, &v)))
+      return error;
+  } else {
+    if ((error = vfs_namelookup(pathname, &v)))
+      return error;
+  }
+
+  if (!error && flags & O_TRUNC) {
+    vattr_t va;
+    vattr_null(&va);
+    va.va_size = 0;
+    error = VOP_SETATTR(v, &va);
+  }
+
+  if (!error)
+    error = VOP_OPEN(v, flags, f);
+
+  /* Drop our reference to v. We received it from vfs_namelookup, but we no
+     longer need it - file f keeps its own reference to v after open. */
+  vnode_drop(v);
+  return error;
+}
 
 int do_open(proc_t *p, char *pathname, int flags, mode_t mode, int *fdp) {
   int error;
@@ -40,58 +133,8 @@ fail:
   return error;
 }
 
-int do_close(proc_t *p, int fd) {
-  return fdtab_close_fd(p->p_fdtable, fd);
-}
-
-int do_read(proc_t *p, int fd, uio_t *uio) {
-  file_t *f;
-  int error;
-
-  if ((error = fdtab_get_file(p->p_fdtable, fd, FF_READ, &f)))
-    return error;
-  error = FOP_READ(f, uio);
-  file_drop(f);
-  return error;
-}
-
-int do_write(proc_t *p, int fd, uio_t *uio) {
-  file_t *f;
-  int error;
-
-  if ((error = fdtab_get_file(p->p_fdtable, fd, FF_WRITE, &f)))
-    return error;
-  error = FOP_WRITE(f, uio);
-  file_drop(f);
-  return error;
-}
-
-int do_lseek(proc_t *p, int fd, off_t offset, int whence, off_t *newoffp) {
-  /* TODO: RW file flag! For now we just file_get_read */
-  file_t *f;
-  int error;
-
-  if ((error = fdtab_get_file(p->p_fdtable, fd, 0, &f)))
-    return error;
-  error = FOP_SEEK(f, offset, whence, newoffp);
-  file_drop(f);
-  return error;
-}
-
-int do_fstat(proc_t *p, int fd, stat_t *sb) {
-  file_t *f;
-  int error;
-
-  if ((error = fdtab_get_file(p->p_fdtable, fd, FF_READ, &f)))
-    return error;
-  error = FOP_STAT(f, sb);
-  file_drop(f);
-  return error;
-}
-
 int do_fstatat(proc_t *p, int fd, char *path, stat_t *sb, int flag) {
-  file_t *f;
-  vnode_t *v, *atdir = NULL;
+  vnode_t *v;
   vattr_t va;
   int error;
   uint32_t vnrflags = 0;
@@ -99,106 +142,13 @@ int do_fstatat(proc_t *p, int fd, char *path, stat_t *sb, int flag) {
   if (!(flag & AT_SYMLINK_NOFOLLOW))
     vnrflags |= VNR_FOLLOW;
 
-  if (fd != AT_FDCWD) {
-    if ((error = fdtab_get_file(p->p_fdtable, fd, FF_READ, &f)))
-      return error;
-    atdir = f->f_vnode;
-  }
-  error = vfs_namelookupat(path, atdir, vnrflags, &v);
-  if (fd != AT_FDCWD)
-    file_drop(f);
-  if (error)
+  if ((error = vfs_namelookupat(p, fd, vnrflags, path, &v)))
     return error;
+
   if (!(error = VOP_GETATTR(v, &va)))
     vattr_convert(&va, sb);
 
   vnode_drop(v);
-  return error;
-}
-
-int do_dup(proc_t *p, int oldfd, int *newfdp) {
-  file_t *f;
-  int error;
-
-  if ((error = fdtab_get_file(p->p_fdtable, oldfd, 0, &f)))
-    return error;
-
-  error = fdtab_install_file(p->p_fdtable, f, 0, newfdp);
-  file_drop(f);
-  return error;
-}
-
-int do_dup2(proc_t *p, int oldfd, int newfd) {
-  file_t *f;
-  int error;
-
-  if (oldfd == newfd)
-    return 0;
-
-  if ((error = fdtab_get_file(p->p_fdtable, oldfd, 0, &f)))
-    return error;
-
-  error = fdtab_install_file_at(p->p_fdtable, f, newfd);
-  file_drop(f);
-  return error;
-}
-
-int do_fcntl(proc_t *p, int fd, int cmd, int arg, int *resp) {
-  file_t *f;
-  int error;
-
-  if ((error = fdtab_get_file(p->p_fdtable, fd, 0, &f)))
-    return error;
-
-  /* TODO: Currently only F_DUPFD command is implemented. */
-  bool cloexec = false;
-  int flags = 0;
-  switch (cmd) {
-    case F_DUPFD_CLOEXEC:
-      cloexec = true;
-      /* FALLTHROUGH */
-    case F_DUPFD:
-      if ((error = fdtab_install_file(p->p_fdtable, f, arg, resp)))
-        break;
-      error = fd_set_cloexec(p->p_fdtable, fd, cloexec);
-      break;
-
-    case F_SETFD:
-      if (arg == FD_CLOEXEC)
-        cloexec = true;
-      error = fd_set_cloexec(p->p_fdtable, fd, cloexec);
-      break;
-
-    case F_GETFD:
-      error = fd_get_cloexec(p->p_fdtable, fd, resp);
-      break;
-
-    case F_GETFL:
-      if (f->f_flags & FF_READ && f->f_flags & FF_WRITE) {
-        flags |= O_RDWR;
-      } else {
-        if (f->f_flags & FF_READ)
-          flags |= O_RDONLY;
-        if (f->f_flags & FF_WRITE)
-          flags |= O_WRONLY;
-      }
-      if (f->f_flags & FF_APPEND)
-        flags |= O_APPEND;
-      *resp = flags;
-      break;
-
-    case F_SETFL:
-      if (arg & O_APPEND)
-        flags |= FF_APPEND;
-      f->f_flags = flags;
-      break;
-
-    default:
-      error = EINVAL;
-      break;
-  }
-
-  file_drop(f);
   return error;
 }
 
@@ -230,46 +180,51 @@ int do_getdents(proc_t *p, int fd, uio_t *uio) {
 }
 
 int do_unlink(proc_t *p, char *path) {
+  vnrstate_t vs;
   int error;
-  vnode_t *vn, *dvn;
-  componentname_t cn;
 
-  if ((error = vfs_namedeleteat(path, NULL, 0, &dvn, &vn, &cn)))
+  if ((error = vnrstate_init(&vs, VNR_DELETE, 0, path)))
     return error;
 
-  if (vn->v_type == V_DIR) {
-    if (dvn == vn)
-      vnode_drop(dvn);
-    else
-      vnode_put(dvn);
-    vnode_put(vn);
-    return EPERM;
-  }
+  if ((error = vfs_nameresolve(&vs)))
+    goto fail;
 
-  error = VOP_REMOVE(dvn, vn, &cn);
-  vnode_put(dvn);
-  vnode_put(vn);
+  if (vs.vs_vp->v_type == V_DIR)
+    error = EPERM;
+  else
+    error = VOP_REMOVE(vs.vs_dvp, vs.vs_vp, &vs.vs_lastcn);
 
+  if (vs.vs_dvp == vs.vs_vp)
+    vnode_drop(vs.vs_dvp);
+  else
+    vnode_put(vs.vs_dvp);
+  vnode_put(vs.vs_vp);
+
+fail:
+  vnrstate_destroy(&vs);
   return error;
 }
 
 int do_mkdir(proc_t *p, char *path, mode_t mode) {
-  int error;
+  vnrstate_t vs;
   vattr_t va;
-  vnode_t *vn, *dvn;
-  componentname_t cn;
+  int error;
 
-  if ((error = vfs_namecreate(path, &dvn, &vn, &cn)))
+  if ((error = vnrstate_init(&vs, VNR_CREATE, VNR_FOLLOW, path)))
     return error;
 
-  if (vn != NULL) {
-    if (vn != dvn)
-      vnode_put(dvn);
-    else
-      vnode_drop(dvn);
+  if ((error = vfs_nameresolve(&vs)))
+    goto fail;
 
-    vnode_drop(vn);
-    return EEXIST;
+  if (vs.vs_vp != NULL) {
+    if (vs.vs_vp != vs.vs_dvp)
+      vnode_put(vs.vs_dvp);
+    else
+      vnode_drop(vs.vs_dvp);
+
+    vnode_drop(vs.vs_vp);
+    error = EEXIST;
+    goto fail;
   }
 
   memset(&va, 0, sizeof(vattr_t));
@@ -278,42 +233,45 @@ int do_mkdir(proc_t *p, char *path, mode_t mode) {
    * https://pubs.opengroup.org/onlinepubs/9699919799/functions/mkdir.html */
   va.va_mode = S_IFDIR | ((mode & ACCESSPERMS) & ~p->p_cmask);
 
-  error = VOP_MKDIR(dvn, &cn, &va, &vn);
+  error = VOP_MKDIR(vs.vs_dvp, &vs.vs_lastcn, &va, &vs.vs_vp);
   if (!error)
-    vnode_drop(vn);
+    vnode_drop(vs.vs_vp);
 
-  vnode_put(dvn);
+  vnode_put(vs.vs_dvp);
+
+fail:
+  vnrstate_destroy(&vs);
   return error;
 }
 
 int do_rmdir(proc_t *p, char *path) {
+  vnrstate_t vs;
   int error;
-  vnode_t *vn, *dvn;
-  componentname_t cn;
 
-  if ((error = vfs_namedelete(path, &dvn, &vn, &cn)))
+  if ((error = vnrstate_init(&vs, VNR_DELETE, VNR_FOLLOW, path)))
     return error;
 
-  if (vn == dvn)
+  if ((error = vfs_nameresolve(&vs)))
+    goto fail;
+
+  if (vs.vs_vp == vs.vs_dvp)
     error = EINVAL;
-  else if (vn->v_type != V_DIR)
+  else if (vs.vs_vp->v_type != V_DIR)
     error = ENOTDIR;
-  else if (vn->v_mountedhere != NULL)
+  else if (vs.vs_vp->v_mountedhere != NULL)
     error = EBUSY;
 
-  if (error) {
-    if (dvn == vn)
-      vnode_drop(dvn);
-    else
-      vnode_put(dvn);
-    vnode_put(vn);
-    return error;
-  }
+  if (!error)
+    error = VOP_RMDIR(vs.vs_dvp, vs.vs_vp, &vs.vs_lastcn);
 
-  error = VOP_RMDIR(dvn, vn, &cn);
-  vnode_put(dvn);
-  vnode_put(vn);
+  if (vs.vs_dvp == vs.vs_vp)
+    vnode_drop(vs.vs_dvp);
+  else
+    vnode_put(vs.vs_dvp);
+  vnode_put(vs.vs_vp);
 
+fail:
+  vnrstate_destroy(&vs);
   return error;
 }
 
@@ -329,19 +287,6 @@ int do_access(proc_t *p, char *path, int amode) {
     return error;
   error = VOP_ACCESS(v, amode);
   vnode_drop(v);
-  return error;
-}
-
-int do_ioctl(proc_t *p, int fd, u_long cmd, void *data) {
-  file_t *f;
-  int error;
-
-  if ((error = fdtab_get_file(p->p_fdtable, fd, 0, &f)))
-    return error;
-  error = FOP_IOCTL(f, cmd, data);
-  file_drop(f);
-  if (error == EPASSTHROUGH)
-    error = ENOTTY;
   return error;
 }
 
@@ -444,19 +389,10 @@ int do_ftruncate(proc_t *p, int fd, off_t length) {
 }
 
 ssize_t do_readlinkat(proc_t *p, int fd, char *path, uio_t *uio) {
-  file_t *f;
-  vnode_t *v, *atdir = NULL;
+  vnode_t *v;
   int error;
 
-  if (fd != AT_FDCWD) {
-    if ((error = fdtab_get_file(p->p_fdtable, fd, FF_READ, &f)))
-      return error;
-    atdir = f->f_vnode;
-  }
-  error = vfs_namelookupat(path, atdir, 0, &v);
-  if (fd != AT_FDCWD)
-    file_drop(f);
-  if (error)
+  if ((error = vfs_namelookupat(p, fd, 0, path, &v)))
     return error;
 
   if (v->v_type != V_LNK)
@@ -466,6 +402,18 @@ ssize_t do_readlinkat(proc_t *p, int fd, char *path, uio_t *uio) {
 
   vnode_drop(v);
   return error;
+}
+
+int do_chdir(proc_t *p, const char *path) {
+  vnode_t *cwd;
+  int error;
+
+  if ((error = vfs_namelookup(path, &cwd)))
+    return error;
+
+  vnode_drop(p->p_cwd);
+  p->p_cwd = cwd;
+  return 0;
 }
 
 int do_fchdir(proc_t *p, int fd) {
@@ -489,45 +437,37 @@ int do_fchdir(proc_t *p, int fd) {
 }
 
 int do_symlinkat(proc_t *p, char *target, int newdirfd, char *linkpath) {
-  file_t *f;
-  vnode_t *vn, *dvn, *atdir = NULL;
-  componentname_t cn;
+  vnrstate_t vs;
   vattr_t va;
   int error;
 
-  if (newdirfd != AT_FDCWD) {
-    if ((error = fdtab_get_file(p->p_fdtable, newdirfd, FF_READ, &f)))
-      return error;
-    atdir = f->f_vnode;
-  }
-  error = vfs_namecreateat(linkpath, atdir, 0, &dvn, &vn, &cn);
-  if (newdirfd != AT_FDCWD)
-    file_drop(f);
-  if (error)
+  if ((error = vnrstate_init(&vs, VNR_CREATE, 0, linkpath)))
     return error;
 
-  if (vn != NULL) {
-    if (vn != dvn)
-      vnode_put(dvn);
-    else
-      vnode_drop(dvn);
+  if ((error = vfs_nameresolveat(p, newdirfd, &vs)))
+    goto fail;
 
-    vnode_drop(vn);
-    return EEXIST;
+  if (vs.vs_vp != NULL) {
+    if (vs.vs_vp != vs.vs_dvp)
+      vnode_put(vs.vs_dvp);
+    else
+      vnode_drop(vs.vs_dvp);
+
+    vnode_drop(vs.vs_vp);
+
+    error = EEXIST;
+    goto fail;
   }
 
   memset(&va, 0, sizeof(vattr_t));
   va.va_mode = S_IFLNK | (ACCESSPERMS & ~p->p_cmask);
 
-  error = VOP_SYMLINK(dvn, &cn, &va, target, &vn);
+  error = VOP_SYMLINK(vs.vs_dvp, &vs.vs_lastcn, &va, target, &vs.vs_vp);
   if (!error)
-    vnode_drop(vn);
-  vnode_put(dvn);
-  return error;
-}
+    vnode_drop(vs.vs_vp);
+  vnode_put(vs.vs_dvp);
 
-int do_umask(proc_t *p, int newmask, int *oldmaskp) {
-  *oldmaskp = p->p_cmask;
-  p->p_cmask = newmask & ALLPERMS;
-  return 0;
+fail:
+  vnrstate_destroy(&vs);
+  return error;
 }
